% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addin.R
\name{gistrAddin}
\alias{gistrAddin}
\title{Create Github gists in RStusio}
\usage{
gistrAddin()
}
\description{
This addin enables you to create Github gists in RStusio.
}
\details{
The addin currently only supports creating single file gists.
More Functionalities of \code{\link{gistr}} package might be added in future
releases.
}

